backend/requirements.txt
# Django core
Django>=4.2,<5.0

# MongoDB backend via Djongo or PyMongo w/ custom auth
djongo>=1.3.6
pymongo>=4.6.0

# JWT Authentication
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3.0

# CORS support
django-cors-headers>=4.3

# For deployment
gunicorn>=21.2.0
python-dotenv>=1.0.1

# Optional: time zone utilities
pytz>=2023.3